+++ examples/demos/event_listeners.py
@@ -53,9 +53,9 @@ node_styles = [
 ]
 
 edge_styles = [
-    EdgeStyle("FOLLOWS", labeled=True, directed=True),
-    EdgeStyle("POSTED", labeled=True, directed=True),
-    EdgeStyle("QUOTES", labeled=True, directed=True),
+    EdgeStyle("FOLLOWS", caption="label", directed=True),
+    EdgeStyle("POSTED", caption="label", directed=True),
+    EdgeStyle("QUOTES", caption="label", directed=True),
 ]
 
 events = [
@@ -71,6 +71,7 @@ with st.container(border=True):
     st.markdown("#### Returned Value")
     st.json(vals or {}, expanded=True)
 
+
 @st.cache_data
 def get_source():
     with open(__file__, "r") as f:
@@ -81,4 +82,3 @@ def get_source():
 source = get_source()
 with st.expander("Source", expanded=False, icon="💻"):
     st.code(source, language="python")
-