+++ st_link_analysis/component/styles.py
@@ -1,4 +1,14 @@
+import warnings
+
 from typing import Optional
+
+
+# TODO: remove if no depreciation warnings
+class LinkAnalysisDeprecationWarning(DeprecationWarning):
+    pass
+
+
+warnings.simplefilter("once", LinkAnalysisDeprecationWarning)
 
 
 class NodeStyle:
@@ -21,7 +31,7 @@ class NodeStyle:
             Specifies the background color of the node. If not provided, the
             default node color "#0a0a0a" will be used.
         caption : Optional[str]
-            Name of the node's attribute to use as caption. If not provided,
+            Name of the node's attribute to use as caption/label. If not provided,
             no caption will be shown.
         icon: Optional[str]
             Node icon to be passed by the name of Material Icons (e.g. 'person')
@@ -61,7 +71,8 @@ class EdgeStyle:
         self,
         label: str,
         color: Optional[str] = None,
-        labeled: bool = False,
+        caption: Optional[str] = None,
+        labeled: bool = False,  # deprecated
         directed: bool = False,
         curve_style: Optional[str] = None,
     ) -> None:
@@ -73,17 +84,21 @@ class EdgeStyle:
         label : str
             The label of the edge. This label is used to identify the group or
             category of the edge.
-        color : Optional[str], optional
+        color : Optional[str]
             Specifies the color of the edge line.
-        labeled : bool, optional
-            Indicates whether the edge should be annotated with its label.
-            Default is False.
-        directed : bool, optional
+        caption : Optional[str], default None
+            Name of the edge's attribute to use as caption/label. If not provided,
+            no caption will be shown.
+        labeled : bool, default False (deprecated)
+            This parameter is deprecated and will be removed in a future release. Use
+            `caption` instead to specify edge caption/label. If `labeled` is set to True
+            and `caption` is not provided, default caption 'label' will be used.
+        directed : bool, default False
             Indicates whether the edge is directed. If True, the edge will be
             rendered with an arrow pointing from the source to target. Default
-            is False. Note: Arrows will not be displayed if `curve_style` 
+            is False. Note: Arrows will not be displayed if `curve_style`
             is set to "haystack".
-        curve_style: bool, optional
+        curve_style: Optional[str]
             Specifies the edge curving method to use. By default, it is set to
             "bezier", which is suitable for multigraphs. For large, simple graphs,
             consider using "haystack" for better performance. For more options
@@ -95,9 +110,19 @@ class EdgeStyle:
         """
         self.label = label
         self.color = color
-        self.labeled = labeled
+        self.caption = caption
         self.directed = directed
         self.curve_style = curve_style
+
+        # TODO: remove in next version along with imports, docs, and signature
+        if labeled is not None:
+            warnings.warn(
+                "Paramter `labeled` is deprecated and will be removed in a future release. "
+                "Please use the `caption` parameter instead.",
+                LinkAnalysisDeprecationWarning,
+            )
+        if labeled and not caption:
+            self.caption = "label"
 
     def dump(self) -> dict:
         selector = f"edge[label='{self.label}']"
@@ -108,13 +133,13 @@ class EdgeStyle:
             style["background-color"] = self.color
             style["text-background-color"] = self.color
             style["target-arrow-color"] = self.color
-        if self.labeled:
-            style["label"] = "data(label)"
+        if self.caption:
+            style["label"] = f"data({self.caption})"
         if self.directed:
             style["target-arrow-shape"] = "triangle"
         if self.curve_style:
             style["curve-style"] = self.curve_style
-        
+
         return {
             "selector": selector,
             "style": style,